Refactor db queries to remove small sql files and place directly in index
