--- src/database/index.js
@@ -1,9 +1,0 @@
-module.exports = {
-  dbSchema: './sql-queries/dbSchema.sql',
-  testData: './sql-queries/testData.sql',
-  getEvents: './sql-queries/getEvents.sql',
-  getSingleEvent: './sql-queries/getSingleEvent.sql',
-  createevent: './sql-queries/createevent.sql',
-  getHashedPassword: './sql-queries/getHashedPassword.sql',
-  getUserDetails: './sql-queries/getUserDetails.sql'
-}
