--- src/database/sql-queries/getEvents.sql
@@ -1,1 +1,0 @@
-SELECT title,short_desc,venue,location, to_char(event_date, 'DD/MM/YYYY') AS event_date,tkts_available FROM events;
