--- src/database/sql-queries/getHashedPassword.sql
@@ -1,1 +1,0 @@
-SELECT password FROM users WHERE email = $1
