--- src/database/sql-queries/getSingleEvent.sql
@@ -1,1 +1,0 @@
-SELECT title, short_desc, long_desc, venue, location, to_char(event_date, 'DD/MM/YYYY') AS event_date, to_char(start_time, 'HH:MM') AS start_time, to_char(end_date, 'DD/MM/YYYY')AS end_date, to_char(end_time, 'HH:MM') as end_time,min_age, tkts_available, info, max_allocation, org_id FROM events WHERE title = $1;
