--- src/database/sql-queries/getUserDetails.sql
@@ -1,1 +1,0 @@
-SELECT id, admin FROM users WHERE email = $1
