+++ src/handlers/home.js
@@ -1,8 +1,8 @@
-const queryDb = require('../database/queryDb')
-const sql = require('../database/index')
+const db = require('../database/dbConnection')
+const sql = require('../database/index-sql')
 
 const getEvents = () => {
-  return queryDb(sql.getEvents)
+  return db.query(sql.getEvents)
     .then(events => {
       return JSON.parse(JSON.stringify(events))
     })
