+++ tests/handlers.home.test.js
@@ -1,7 +1,5 @@
 const tape = require('tape')
 const { getEvents } = require('../src/handlers/home.js')
-const queryDb = require('../src/database/queryDb.js')
-const sql = require('../src/database/index.js')
 const resetTestDb = require('../src/database/resetTestDb')
 
 const expected = [ { title: 'Fac Welcome',
